various renamings and directory shuffling
