+++ src/interpreter/execute.cpp
@@ -4,15 +4,15 @@
 #include "../ast_allocator.hpp"
 #include "../compute_offsets.hpp"
 #include "../cst_allocator.hpp"
-#include "../ct_eval.hpp"
 #include "../lexer.hpp"
 #include "../match_identifiers.hpp"
-#include "../metacheck.hpp"
 #include "../parser.hpp"
 #include "../symbol_table.hpp"
 #include "../token_array.hpp"
-#include "../typecheck.hpp"
-#include "../typechecker.hpp"
+#include "../typechecker/ct_eval.hpp"
+#include "../typechecker/metacheck.hpp"
+#include "../typechecker/typecheck.hpp"
+#include "../typechecker/typechecker.hpp"
 #include "eval.hpp"
 #include "garbage_collector.hpp"
 #include "interpreter.hpp"
